Add enhanced user popover and create new Text component
Popover content has been added inside the CurrentUser component, to show additional user information as well as a button which will toggle an upcoming account settings modal. In addition, a new Text component was created, incorporating a range of sizes and styles. This new component has been used to display the user's name in the popover content.