UpcomingEvents: Add data loading state and data ui markup (w placeholders)
Added a loading state and related UI to the UpcomingEvents component, with a conditional render based on the isLoading state. By monitoring this state variable, we can show a loading list first and switch to the actual content when it's ready. For now, it shows dummy data, but will show actual data (or loading skeletons) as we continue development.