Improve UI for loaded upcoming events and add test data
This commit modifies the dashboard’s upcoming-events component to include antd’s Badge component. Dummy data is added to the dataSource array in the List component, replacing the previous placeholder array of numbers. The loading state is also now initially set to false, as we are building the loaded state UI.