Add dayjs library and filter, sort and paginate upcoming events
The dayjs library has been added to the package.json and package-lock.json files as a dependency for handling dates. In the upcoming-events.tsx file, pagination, sorters, and filters have been added; the filters utilize the dayjs library to only include events that have a start date of today or later. The list of upcoming events is also now sorted in ascending order based on start date.